+++ util/include/Options.h
@@ -1,4 +1,4 @@
-/// 20-May-2010 WGS 
+/// 20-May-2020 WGS 
 /// Implement a program-wide job-options processing service.
 
 /// See README.md for documentation.
@@ -6,6 +6,7 @@
 #ifndef Options_h
 #define Options_h 1
 
+#include <TDirectory.h>
 #include <map>
 #include <string>
 
@@ -24,11 +25,14 @@ namespace util {
       return &instance;
     }
 
-    /// Initialize the options from an XML file.  This should be called
-    /// just once, from the main routine.  Note: This routine uses
-    /// getopt_long, a standard GNU utility that will "mangle" the
-    /// contents of argv.
-    bool ParseOptions(int argc, char** argv, const std::string& programName);
+    /// ParseOptions initializes the options from an XML file or a
+    /// ROOT file. This should be called just once, from the main
+    /// routine. 
+
+    /// Note: This routine uses getopt_long, a standard GNU utility
+    /// that will "mangle" the contents of argv.
+    bool ParseOptions(int argc, char** argv, 
+		      const std::string programName = "");
     
     /// The "getters", one for each type of value.
     bool GetOption(const std::string name, double& value) const;
@@ -52,28 +56,39 @@ namespace util {
     std::string GetOptionType( size_t i ) const;
     std::string GetOptionBrief( size_t i ) const;
     std::string GetOptionDescription( size_t i ) const;
+    std::string GetOptionSource( size_t i ) const;
+
+    // The argument to this method is the output directory for the
+    // ntuple. Presumably this will be a ROOT file.
+    bool WriteNtuple(TDirectory* output, std::string outputNtuple = "Options");
+
 
   protected:
     /// Standard null constructor for a singleton class.
     Options() {}
     
   private:
-    enum m_option_type { e_string, e_double, e_integer, e_boolean };
+    enum m_option_type { e_string, e_double, e_integer, e_boolean, e_flag};
     struct m_option_attributes {
       std::string value;   ///< The option's value
-      m_option_type type;  ///< string/double/integer/boolean
+      m_option_type type;  ///< string/double/integer/boolean/flag
       char brief;          ///< Optional one-char short option
       std::string desc;    ///< Optional description for --help
+      std::string source;  ///< From which place (XML tag block, command line) did this come from?
     };
 
     std::map<std::string,m_option_attributes> m_options; ///< User options map
 
     // Parse the contents of the XML file... once we've
     // determined the name of that file. 
-    bool m_ParseXML(const std::string filename, const std::string programTag);
+    bool m_ParseXML(const std::string& filename, const std::string& programTag);
 
-    std::string m_progName;    ///< The name of the running program (argv[0])
-    std::string m_xmlFile;     ///< The name of the options XML file used
+    // If it turns out that the options file is a ROOT file, look for
+    // an Options ntuple.
+    bool m_RootOptions(const std::string& filename);
+
+    std::string m_progPath;    ///< The path of the running program (argv[0])
+    std::string m_optionsFile; ///< The name of the options XML file used
   };
 
 } // namespace util